+++ zellij-server/src/tab/unit/tab_integration_tests.rs
@@ -177,7 +177,7 @@ impl MockPtyInstructionBus {
                         .recv()
                         .expect("failed to receive event on channel");
                     match event {
-                        PtyWriteInstruction::Write(msg, _) => output
+                        PtyWriteInstruction::Write(msg, _, _) => output
                             .lock()
                             .unwrap()
                             .push(String::from_utf8_lossy(&msg).to_string()),