+++ zellij-utils/assets/prost_ipc/client_server_contract.rs
@@ -119,7 +119,7 @@ pub struct RgbColor {
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Action {
-    #[prost(oneof="action::ActionType", tags="1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 98, 87, 88, 89, 95, 96, 90, 91, 92, 93, 94, 97")]
+    #[prost(oneof="action::ActionType", tags="1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100")]
     pub action_type: ::core::option::Option<action::ActionType>,
 }
 /// Nested message and enum types in `Action`.
@@ -299,18 +299,12 @@ pub mod action {
         KeybindPipe(super::KeybindPipeAction),
         #[prost(message, tag="86")]
         ListClients(super::ListClientsAction),
-        #[prost(message, tag="98")]
-        ListPanes(super::ListPanesAction),
         #[prost(message, tag="87")]
         TogglePanePinned(super::TogglePanePinnedAction),
         #[prost(message, tag="88")]
         StackPanes(super::StackPanesAction),
         #[prost(message, tag="89")]
         ChangeFloatingPaneCoordinates(super::ChangeFloatingPaneCoordinatesAction),
-        #[prost(message, tag="95")]
-        TogglePaneBorderless(super::TogglePaneBorderlessAction),
-        #[prost(message, tag="96")]
-        SetPaneBorderless(super::SetPaneBorderlessAction),
         #[prost(message, tag="90")]
         TogglePaneInGroup(super::TogglePaneInGroupAction),
         #[prost(message, tag="91")]
@@ -321,8 +315,18 @@ pub mod action {
         NewBlockingPane(super::NewBlockingPaneAction),
         #[prost(message, tag="94")]
         OverrideLayout(super::OverrideLayoutAction),
+        #[prost(message, tag="95")]
+        TogglePaneBorderless(super::TogglePaneBorderlessAction),
+        #[prost(message, tag="96")]
+        SetPaneBorderless(super::SetPaneBorderlessAction),
         #[prost(message, tag="97")]
         SaveSession(super::SaveSessionAction),
+        #[prost(message, tag="98")]
+        ListPanes(super::ListPanesAction),
+        #[prost(message, tag="99")]
+        WriteCharsToPaneId(super::WriteCharsToPaneIdAction),
+        #[prost(message, tag="100")]
+        WriteToPaneId(super::WriteToPaneIdAction),
     }
 }
 // Action message definitions (all 92 variants)
@@ -567,6 +571,22 @@ pub struct WriteAction {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct WriteCharsAction {
     #[prost(string, tag="1")]
+    pub chars: ::prost::alloc::string::String,
+}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct WriteToPaneIdAction {
+    #[prost(message, optional, tag="1")]
+    pub pane_id: ::core::option::Option<PaneId>,
+    #[prost(uint32, repeated, tag="2")]
+    pub bytes: ::prost::alloc::vec::Vec<u32>,
+}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct WriteCharsToPaneIdAction {
+    #[prost(message, optional, tag="1")]
+    pub pane_id: ::core::option::Option<PaneId>,
+    #[prost(string, tag="2")]
     pub chars: ::prost::alloc::string::String,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]