+++ zellij-utils/src/errors.rs
@@ -364,6 +364,7 @@ pub enum ScreenContext {
     ResizePaneWithId,
     EditScrollbackForPaneWithId,
     WriteToPaneId,
+    WriteKeyToPaneId,
     CopyTextToClipboard,
     MovePaneWithPaneId,
     MovePaneWithPaneIdInDirection,